Integration ws compiler with express server api
