--- api/compiler/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "subserver",
-  "version": "1.0.0",
-  "description": "bella li",
-  "main": "index.js",
-  "scripts": {
-    "start": "node subserver.js"
-  },
-  "author": "Porcodio",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "dockerode": "^3.3.1",
-    "fs": "^0.0.1-security",
-    "https": "^1.0.0",
-    "node": "^17.7.2",
-    "uuid": "^8.3.2",
-    "websocket": "^1.0.34",
-    "ws": "^8.6.0"
-  },
-  "type": "module"
-}
