+++ package.json
@@ -26,15 +26,20 @@
   ],
   "dependencies": {
     "cors": "^2.8.5",
+    "dockerode": "^3.3.1",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "express-jwt": "^7.6.2",
+    "fs": "^0.0.1-security",
     "jwks-rsa": "^2.1.0",
     "mongodb": "^4.4.1",
     "mongoose": "^6.2.9",
     "node-fetch": "2",
+    "query-string": "^7.1.1",
+    "stream": "^0.0.2",
     "underscore": "^1.13.3",
-    "uuid": "^8.3.2"
+    "uuid": "^8.3.2",
+    "ws": "^8.6.0"
   },
   "engines": {
     "node": ">=16.0"
